admin_service/requirements/base.txt
django==3.1
django-model-utils==4.0.0  # Набор полезных базовых классов и утилит для Django
django-extensions==3.1.1
django-debug-toolbar==3.2
psycopg2-binary==2.8.6